Move helper methods to separate src file
